- Refactores the code to pre-define the size of lists and maps when instantiating them to improve performance. - Updates the code to load classes from the new CloudSim Plus package structure. - Adds EditorConfig configuration file.

+++ src/main/java/cloudreports/models/HostRegistry.java
@@ -1,13 +1,13 @@
-/* 
+/*
  * Copyright (c) 2010-2012 Thiago T. Sá
- * 
+ *
  * This file is part of CloudReports.
  *
  * CloudReports is free software: you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
  * the Free Software Foundation, either version 3 of the License, or
  * (at your option) any later version.
- * 
+ *
  * CloudReports is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
@@ -19,66 +19,74 @@
 
 package cloudreports.models;
 
+import org.cloudbus.cloudsim.power.models.PowerModel;
+import org.cloudbus.cloudsim.provisioners.PeProvisioner;
+import org.cloudbus.cloudsim.provisioners.ResourceProvisioner;
+import org.cloudbus.cloudsim.resources.Bandwidth;
+import org.cloudbus.cloudsim.resources.FileStorage;
+import org.cloudbus.cloudsim.resources.Pe;
+import org.cloudbus.cloudsim.resources.Ram;
+import org.cloudbus.cloudsim.schedulers.vm.VmScheduler;
+
 import java.io.Serializable;
 
 /**
  * A host registry stores information about a specific host configuration.
  * It contains general information such as scheduling policy, power specifications
  * and amount of resources.
- * 
+ *
  * @author      Thiago T. Sá
  * @since       1.0
  */
 public final class HostRegistry implements Serializable{
-
     /** The host's id. */
-    private long id;
-    
-    /** The host's scheduling policy. */
+    private int id;
+
+    /** The name suffix of the host's {@link VmScheduler} class. */
     private String schedulingPolicyAlias;
-    
-    /** The host's number of processing elements. */
+
+    /** The host's number of {@link Pe}. */
     private int numOfPes;
-    
-    /** The amount of mips per processing units. */
+
+    /** The amount of mips by {@link Pe}. */
     private double mipsPerPe;
-    
+
     /** The maximum power consumption. */
     private double maxPower;
-    
+
     /** The static power consumption percent. */
     private double staticPowerPercent;
-    
-    /** The host's power model. */
+
+    /** The host's {@link PowerModel}. */
     private String powerModelAlias;
-    
-    /** The amount of RAM. */
+
+    /** The amount of {@link Ram}. */
     private int ram;
-    
-    /** The host's RAM provisioner. */
+
+    /** The name suffix of the host's RAM {@link ResourceProvisioner} class. */
     private String ramProvisionerAlias;
-    
-    /** The amount of bandwidth. */
+
+    /** The amount of {@link Bandwidth}. */
     private long bw;
-    
-    /** The host's bandwidth provisioner. */
+
+    /** The name suffix of the host's {@link Bandwidth} {@link ResourceProvisioner} class. */
     private String bwProvisionerAlias;
-    
+
     /** The amount of hosts with this registry's specification
      * owned by the datacenter. */
     private int amount;
-    
-    /** The storage capacity. */
+
+    /** The {@link FileStorage Storage} capacity. */
     private long storage;
-    
-    /** The processing elements provisioner. */
+
+    /** The name suffix of the host's {@link PeProvisioner} class. */
     private String peProvisionerAlias;
 
-    /** 
+    /**
      * Creates a new host registry.
-     * 
-     * @since           1.0
-     */    
+     *
+     * @since 1.0
+     */
     public HostRegistry() {
         setAmount(1);
         setSchedulingPolicyAlias("Time shared");
@@ -98,25 +106,25 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Gets the host's id.
-     * 
+     *
      * @return the host's id.
      */
-    public long getId() {
+    public int getId() {
         return id;
     }
 
     /**
      * Sets the host's id.
-     * 
+     *
      * @param   id  the host's id.
      */
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
 
     /**
      * Gets the host's scheduling policy.
-     * 
+     *
      * @return the host's scheduling policy.
      */
     public String getSchedulingPolicyAlias() {
@@ -125,7 +133,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Sets the host's scheduling policy.
-     * 
+     *
      * @param   schedulingPolicyAlias   the host's scheduling policy.
      */
     public void setSchedulingPolicyAlias(String schedulingPolicyAlias) {
@@ -134,7 +142,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Gets the host's number of processing elements.
-     * 
+     *
      * @return the host's number of processing elements.
      */
     public int getNumOfPes() {
@@ -143,7 +151,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Sets the host's number of processing elements.
-     * 
+     *
      * @param   numOfpes    the host's number of processing elements.
      */
     public void setNumOfPes(int numOfpes) {
@@ -152,7 +160,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Gets the amount of mips per processing elements.
-     * 
+     *
      * @return the amount of mips per processing elements.
      */
     public double getMipsPerPe() {
@@ -161,7 +169,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Sets the amount of mips per processing elements.
-     * 
+     *
      * @param   mipsPerPe   the amount of mips per processing elements.
      */
     public void setMipsPerPe(double mipsPerPe) {
@@ -170,7 +178,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Gets the host's maximum power consumption.
-     * 
+     *
      * @return the host's maximum power consumption.
      */
     public double getMaxPower() {
@@ -179,7 +187,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Sets the host's maximum power consumption.
-     * 
+     *
      * @param   maxPower    the host's maximum power consumption.
      */
     public void setMaxPower(double maxPower) {
@@ -188,7 +196,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Gets the host's static power consumption percent.
-     * 
+     *
      * @return the host's static power consumption percent.
      */
     public double getStaticPowerPercent() {
@@ -196,7 +204,7 @@ public final class HostRegistry implements Serializable{
     }
     /**
      * Sets the host's static power consumption percent.
-     * 
+     *
      * @param   staticPowerPercent  the host's static power consumption percent.
      */
     public void setStaticPowerPercent(double staticPowerPercent) {
@@ -205,7 +213,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Gets the host's power model alias.
-     * 
+     *
      * @return the host's power model alias.
      */
     public String getPowerModelAlias() {
@@ -214,7 +222,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Sets the host's power model alias.
-     * 
+     *
      * @param   powerModelAlias the host's power model alias.
      */
     public void setPowerModelAlias(String powerModelAlias) {
@@ -223,7 +231,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Gets the host's amount of RAM.
-     * 
+     *
      * @return the host's amount of RAM.
      */
     public int getRam() {
@@ -232,7 +240,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Sets the host's amount of RAM.
-     * 
+     *
      * @param   ram the host's amount of RAM.
      */
     public void setRam(int ram) {
@@ -241,7 +249,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Gets the host's RAM provisioner.
-     * 
+     *
      * @return the host's RAM provisioner.
      */
     public String getRamProvisionerAlias() {
@@ -250,7 +258,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Sets the host's RAM provisioner.
-     * 
+     *
      * @param   ramProvisionerAlias the host's RAM provisioner.
      */
     public void setRamProvisionerAlias(String ramProvisionerAlias) {
@@ -259,7 +267,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Gets the host's bandwidth.
-     * 
+     *
      * @return the host's bandwidth.
      */
     public long getBw() {
@@ -268,7 +276,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Sets the host's bandwidth.
-     * 
+     *
      * @param   bw  the host's bandwidth.
      */
     public void setBw(long bw) {
@@ -277,7 +285,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Gets the host's bandwidth provisioner.
-     * 
+     *
      * @return the host's bandwidth provisioner.
      */
     public String getBwProvisionerAlias() {
@@ -286,7 +294,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Sets the host's bandwidth provisioner.
-     * 
+     *
      * @param bwProvisionerAlias    the host's bandwidth provisioner.
      */
     public void setBwProvisionerAlias(String bwProvisionerAlias) {
@@ -295,7 +303,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Gets the amount of hosts with this configuration.
-     * 
+     *
      * @return the amount of hosts with this configuration.
      */
     public int getAmount() {
@@ -304,7 +312,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Sets the amount of hosts with this configuration.
-     * 
+     *
      * @param   amount  the amount of hosts with this configuration.
      */
     public void setAmount(int amount) {
@@ -313,7 +321,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Gets the host's storage capacity.
-     * 
+     *
      * @return the host's storage capacity.
      */
     public long getStorage() {
@@ -322,7 +330,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Sets the host's storage capacity.
-     * 
+     *
      * @param   storage the host's storage capacity.
      */
     public void setStorage(long storage) {
@@ -331,7 +339,7 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Gets the host's processing elements provisioner.
-     * 
+     *
      * @return the host's processing elements provisioner.
      */
     public String getPeProvisionerAlias() {
@@ -340,21 +348,21 @@ public final class HostRegistry implements Serializable{
 
     /**
      * Sets the host's processing elements provisioner.
-     * 
+     *
      * @param   peProvisionerAlias  the host's processing elements provisioner.
      */
     public void setPeProvisionerAlias(String peProvisionerAlias) {
         this.peProvisionerAlias = peProvisionerAlias;
     }
-    
+
     /**
      * Indicates whether the host can allocate a virtual machine or not.
-     * 
+     *
      * @param   vmr     the virtual machine to be allocated.
      * @return          <code>true</code> if the host can allocate the virtual
      *                  machine; <code>false</code> otherwise.
      * @since           1.0
-     */    
+     */
     public boolean canRunVM(VirtualMachineRegistry vmr) {
         if(this.getRam()<vmr.getRam()) return false;
         if((this.getNumOfPes()*this.getMipsPerPe()) < vmr.getMips()) return false;
@@ -363,7 +371,7 @@ public final class HostRegistry implements Serializable{
 
         return true;
     }
-    
+
     @Override
     public boolean equals(Object host){
       if ( this == host ) return true;
@@ -377,7 +385,7 @@ public final class HostRegistry implements Serializable{
         int hash = 3;
         hash = 97 * hash + (int) (this.id ^ (this.id >>> 32));
         return hash;
-    }    
+    }
 
     @Override
     public String toString() {
